requirements.txt
gunicorn
Flask
numpy
Pillow
tensorflow